QuantizationMetadata class
