--- src/compressed_tensors/quantization/quant_names.py
@@ -1,43 +1,0 @@
-# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from enum import Enum
-
-
-__all__ = ["ALL_QPARAM_NAMES", "KVCacheScaleType"]
-
-
-class KVCacheScaleType(Enum):
-    KEY = "k_scale"
-    VALUE = "v_scale"
-
-
-ALL_QPARAM_NAMES = (
-    [KVCacheScaleType.KEY.value, KVCacheScaleType.VALUE.value]
-    + [
-        f"{base_name}_{suffix}"
-        for base_name in ("input", "weight", "output")
-        for suffix in (
-            "global_scale",
-            "scale",
-            "zero_point",
-            "g_idx",
-        )
-    ]
-    + [
-        "weight_packed",
-        "weight_global_scale",
-        "weight_shape",
-    ]
-)
